requirements.txt
Flask==2.0.2
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
psycopg2==2.9.1
python-dotenv==0.19.1
SQLAlchemy==1.4.26
Werkzeug==2.0.2
